pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "adaptive-english-ai-tutor"
version = "0.1.0"
description = "Adaptive English AI Tutor prototype for offline tutoring experiments."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.7,<3",
  "PyYAML>=6.0,<7",
  "openai>=1.40,<2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9",
]
ml = [
  "torch>=2.0,<3",
]

[project.scripts]
run-experiment = "adaptive_tutor.cli:run_experiment_entry"
generate-report = "adaptive_tutor.cli:generate_report_entry"
train-reading-triage = "adaptive_tutor.cli:train_reading_triage_entry"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["src"]
